requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.3
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.9
dotenv==0.9.9
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fsspec==2025.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==5.30.0
gradio_client==1.10.1
groovy==0.1.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.6
jupyter_client==8.6.2
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.2.6
orjson==3.10.18
packaging==24.0
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
pipenv==2025.0.2
platformdirs==4.2.2
prompt-toolkit==3.0.43
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.8
psycopg2==2.9.10
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
requests==2.32.3
rich==14.0.0
rsa==4.9.1
ruff==0.11.10
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
tomlkit==0.13.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.20.0
virtualenv==20.30.0
wcwidth==0.2.13
websockets==15.0.1
